requirements.txt
black
gdown
isort
scikit-learn
pandas
lpips==0.1.4
monai-generative
monai[nibabel, matplotlib, pandas, mlflow, tqdm, pydicom]==1.2.0
numpy==1.22.2
omegaconf==2.3.0
pre-commit
protobuf==3.20.3
pynvml
spacy
tensorboardX
wandb
tifffile
torchmetrics
torch-fidelity
pydantic==1.10.11